moved source as suggested; tidied up some repetition
